refactor: rounds-list component structure
