+++ src/components/rounds-list/index.ts
@@ -1,2 +1,7 @@
 export * from "./rounds-list"
-export * from "./name"
+export * from "./round-title"
+export * from "./round-item"
+export * from "./participants-list"
+export * from "./participant-item"
+export * from "./participant-image"
+export * from "./participant-name"
