--- src/components/rounds-list/name/index.ts
@@ -1,1 +1,0 @@
-export * from "./name"
