+++ src/components/rounds-list/participant-image.tsx
