+++ src/components/rounds-list/round-item.tsx
@@ -1,0 +1,5 @@
+import { PropsWithChildren } from "react"
+
+export function RoundItem({ children }: PropsWithChildren) {
+  return <div className="mt-8 md:mt-0">{children}</div>
+}
