+++ src/shared/types/index.ts
@@ -1,2 +1,3 @@
 export * from "./participant"
 export * from "./round"
+export * from "./event"
